Version canidate, Enough features work for v0.1.0
WKami

+++ commandwrapper.py
@@ -3,14 +3,17 @@
 
 import os
 import sys
-import getopt
+
 cogfile = 'cogs/commands.py'
 emptycog = 'templates/emptycog.py'
 
 def buildcog():
     #build cog
     with open('commandlist.txt', 'r') as com:
-        os.remove(cogfile)
+        #check if cog file exists, is so delete it
+        if os.path.isfile(cogfile) == True:
+            os.remove(cogfile)
+        #copy empty cog to cog file
         with open(cogfile, 'w') as f:
             with open(emptycog, 'r') as e:
                 f.write(e.read())
@@ -26,7 +29,7 @@ def buildcog():
                 #break line into command, response, responsefile, volume
                 command, response, responsefile, volume = line.split(':')
                 #insert command into cog file at index, increment index
-                lines.insert(index, "@commands.command()\n")
+                lines.insert(index, "    @commands.command()\n")
                 index += 1
                 lines.insert(index, "    async def " + command + "(self, ctx: commands.Context):\n")
                 index += 1
@@ -34,10 +37,18 @@ def buildcog():
                     lines.insert(index, "        await ctx.send('" + response + "')\n")
                     index += 1
                 elif response == "none":
-                    lines.insert(index, "        self.player.play(ctx, '" + responsefile + "', volume=" + volume + ")\n")
+                    lines.insert(index, "        self.player.volume = " + volume + "\n")
+                    index += 1
+                    lines.insert(index, "        psound = sounds.Sound(source='" + responsefilePath + responsefile +"')\n")
+                    index += 1
+                    lines.insert(index, "        self.player.play(psound)\n")
                     index += 1
                 else:
-                    lines.insert(index, "        self.player.play(ctx, '" + responsefile + "', volume=" + volume + ")\n")
+                    lines.insert(index, "        self.player.volume = " + volume + "\n")
+                    index += 1
+                    lines.insert(index, "        psound = sounds.Sound(source='" + responsefilePath + responsefile +"')\n")
+                    index += 1
+                    lines.insert(index, "        self.player.play(psound)\n")
                     index += 1
                     lines.insert(index, "        await ctx.send('" + response + "')\n")
                     index += 1
@@ -55,7 +66,7 @@ responsefilePath = 'sounds/'
 
 
 #get arguments
-opts, args = getopt.getopt(sys.argv[1:], "AR:c:r:f:v:")
+
 #opts is a list of tuples, each tuple is of the form (flag, argument)
 #args is a list of arguments that are not flags
 
@@ -73,20 +84,36 @@ response = "none"
 responsefile = "none"
 volume = "100"
 
-#set flags
-for opt, arg in opts:
-    if opt == "-A":
+#set flags,
+if len(sys.argv) > 1:
+    args = sys.argv[1:]
+    print(args)
+    #check if -A or -R is present
+    if '-A' in args:
         add = True
-    elif opt == "-R":
+    elif '-R' in args:
         remove = True
-    elif opt == "-c":
-        command = arg
-    elif opt == "-r":
-        response = arg
-    elif opt == "-f":
-        responsefile = arg
-    elif opt == "-v":
-        volume = arg
+    else:
+        print("Must specify -A or -R")
+        exit(1)
+    #check if -c is present
+    if '-c' in args:
+        command = args[args.index('-c') + 1]
+    else:
+        print("Must specify command")
+        exit(1)
+    #check if -r is present
+    if '-r' in args:
+        response = args[args.index('-r') + 1]
+    #check if -f is present
+    if '-f' in args:
+        responsefile = args[args.index('-f') + 1]
+    #check if -v is present
+    if '-v' in args:
+        volume = args[args.index('-v') + 1]
+
+
+
 
 #check if arguments are valid
 if add == remove:
@@ -95,32 +122,41 @@ if add == remove:
 if command == "none":
     print("Must specify command")
     exit(1)
-if response == "none" and responsefile == "none":
-    print("Must specify response or responsefile")
-    exit(1)
-if responsefile != "none" and not os.path.exists(responsefile):
-    print("Responsefile not found, must be in sounds directory")
-    exit(1)
+#oly check the following if adding a command
+if add == True:
+    if responsefile != "none":
+        if os.path.isfile(responsefilePath + responsefile) == False:
+            print("Response file not found")
+            exit(1)
+        elif responsefile[-4:] != ".mp3":
+            print("Response file must be an mp3")
+            exit(1)
+    if response == "none" and responsefile == "none":
+        print("Must specify response or responsefile")
+        exit(1)
+    if int(volume) < 0 or int(volume) > 100:
+        print("Volume must be between 0 and 100")
+        exit(1)
 
 #check if commandlist.txt exists, if not create it comands will be stored in commandlist.txt in the format "command:response:responsefile:volume"
 if not os.path.exists('commandlist.txt'):
     open('commandlist.txt', 'w').close()
 
-commands = []
+exist = []
 #read existing commands
 with open('commandlist.txt', 'r') as f:
     for line in f:
         splitline = line.split(':')
-        commands.append(splitline[0])
+        exist.append(splitline[0])
 
 #add command
 if add == True:
-    if command in commands:
+    if command in exist:
         print("Command already exists")
         exit(1)
     else:
         with open('commandlist.txt', 'a') as f:
-            f.write(command + ':' + response + ':' + responsefile + ':' + volume + '/n')
+            f.write(command + ':' + response + ':' + responsefile + ':' + volume +'\n')
         if buildcog() == 0:
             print("Command added")
             exit(0)
@@ -130,7 +166,7 @@ if add == True:
 
 #remove command
 elif remove == True:
-    if command in commands:
+    if command in exist:
         with open('commandlist.txt', 'r') as f:
             lines = f.readlines()
         with open('commandlist.txt', 'w') as f:
